+++ packages/idstool/prepare.ts
@@ -3,6 +3,7 @@ import path from "path"
 import Database from "better-sqlite3"
 import { transactionSync } from "./lib/dbutils"
 import { IDSDecomposer } from "./lib/ids-decomposer"
+import { tokenizeIDS } from "./lib/ids-tokenizer"
 
 const mojidb = require.resolve("@mandel59/mojidata/dist/moji.db")
 
@@ -15,6 +16,7 @@ const symbols_in_ids = new Set<string>()
 for (const ids of db.prepare<[], ["IDS"], { IDS: string }>(`SELECT IDS from moji.ids`).pluck().iterate() as Iterable<string>) {
     ids.match(/[\p{Sm}\p{So}\p{Po}]/gu)?.forEach(c => symbols_in_ids.add(c))
 }
+const usource = db.prepare<[], ["U_source_ID", "IDS"], { U_source_ID: string, IDS: string }>(`SELECT U_source_ID, IDS FROM moji.usource WHERE IDS is not null`).all()
 db.exec(`DETACH DATABASE moji`)
 
 db.exec(`drop table if exists "idsfind"`)
@@ -46,12 +48,25 @@ function showProgressForEach<T>(array: T[], proc: (value: T) => void) {
     }
 }
 
-const allCharSources = decomposer.allCharSources()
+const allCharSources: {
+    char: string;
+    IDS?: string;
+    source: string;
+}[] = [
+        ...decomposer.allCharSources(),
+        ...usource.map(({ U_source_ID, IDS }) => ({
+            char: `&${U_source_ID};`,
+            IDS,
+            source: 'U',
+        })),
+    ]
 transactionSync(db, () => {
     const n = allCharSources.length
     console.log("total", n)
-    showProgressForEach(allCharSources, ({ char, source }) => {
-        const alltokens = decomposer.decomposeAll(char, source)
+    showProgressForEach(allCharSources, ({ char, IDS, source }) => {
+        const alltokens = IDS
+            ? decomposer.decomposeTokens(tokenizeIDS(IDS), source)
+            : decomposer.decomposeAll(char, source)
         for (const tokens of alltokens) {
             insert_idsfind.run({ ucs: char, tokens: tokens.join(' ') })
         }
@@ -70,12 +85,20 @@ transactionSync(db, () => {
 db.exec(`INSERT INTO idsfind (UCS, IDS_tokens) SELECT DISTINCT UCS, IDS_tokens FROM idsfind_temp`)
 
 db.exec(`drop table if exists "idsfind_fts"`)
+db.exec(`CREATE TABLE "idsfind_ref" (docid INTEGER PRIMARY KEY, char TEXT NOT NULL)`)
+db.exec(`CREATE INDEX "idsfind_ref_char" ON "idsfind_ref" (char)`)
+db.exec(`INSERT INTO idsfind_ref (docid, char) SELECT rowid, UCS FROM idsfind`)
 db.exec(`CREATE VIRTUAL TABLE "idsfind_fts" USING fts4 (
     content="",
     tokenize=unicode61 "tokenchars={}&-;§${Array.from(symbols_in_ids).join("")}",
     "IDS_tokens"
 )`)
-db.exec(`INSERT INTO idsfind_fts (docid, IDS_tokens) SELECT unicode(UCS) AS docid, '§ ' || group_concat(IDS_tokens, ' § ') || ' §' FROM idsfind WHERE length(UCS) = 1 GROUP BY unicode(UCS)`)
+db.exec(`INSERT INTO idsfind_fts (docid, IDS_tokens) SELECT
+    (SELECT docid FROM idsfind_ref WHERE char = UCS) AS docid,
+    '§ ' || group_concat(IDS_tokens, ' § ') || ' §'
+FROM idsfind
+GROUP BY UCS`)
+db.exec(`DROP INDEX "idsfind_ref_char"`)
 
 db.exec(`PRAGMA journal_mode = delete`)
 db.exec(`PRAGMA page_size = 1024`)